requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
cassandra-driver==3.20.2
certifi==2016.2.28
click==6.7
cycler==0.10.0
Flask==0.12.2
html5lib==0.9999999
numpy==1.12.1
opencv-contrib-python==4.1.2.30
itsdangerous==0.24
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.2
protobuf==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.5
virtualenv==15.1.0
virtualenv-clone==0.4.0
Werkzeug==0.12.2
